requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-mssql==1.8
django-mssql-backend==2.8.1
django-ninja==0.13.2
django-pyodbc-azure==2.1.0.0
ecdsa==0.17.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyodbc==4.0.30
python-jose==3.3.0
pytz==2021.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-jenkins
pep8
pyflakes
pylint
pytest-django
pytest-xdist
